c: player structure refactoring
