+++ client/src/player/canvas.ts
