+++ client/src/player/index.ts
